requirements.txt
python-dotenv==1.0.1
loguru==0.7.3
SQLAlchemy==2.0.40
aiosqlite==0.21.0
greenlet==3.0.3
web3==6.20.4
fake-useragent==2.2.0
aiohttp-socks==0.8.4
curl_cffi==0.11.3
aiohttp==3.11.8
cryptography==44.0.2
aiohttp_proxy==0.1.2
py-solc-x==2.0.3
inquirer~=3.4.0
rich~=13.9.4
colorama~=0.4.6
Faker~=30.4
beautifulsoup4==4.13.4
better-proxy==1.3.1
pyotp==2.9.0
pydantic==2.11.7
tenacity==9.1.2
yarl==1.18.3
ruamel.yaml==0.18.14
pyaml==25.7.0
lxml==6.0.0